move function to access arrays by time to extra module
